Changes to health, enemies and add spawning
HP is displayed on the top right of the screen and can now regenerate after a certain amount of time of not taking damage. Enemy X and Z rotation are now locked and enemies cannot be pushed by player character. To attack, enemies must now remain within attack range for a certain amount of time before damaging the player. Finally, a basic spawning code was added which is used to spawn in at random either running or walking enemies. There is manual spawning which can be activate by the player by pressing P on the keyboard. There is also automatic spawning which will spawn an enemy every few seconds. For this to work, the autospawning option must be checked in the script which can be found on the spawner object. When this is checked, manual spawning is disabled